+++ .github/workflows/presubmit.yml
@@ -30,7 +30,7 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Check format and autogenerated files
-      run: DRY_RUN=1 make regen
+      run: CHECK=1 make regen
   test:
     runs-on: ubuntu-latest
     steps:
@@ -45,5 +45,5 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v3
-    - name: Run kindtest
-      run: make kindtest
+    - name: Run e2e
+      run: make e2e
