+++ Makefile
@@ -36,14 +36,17 @@ define docker_tag_push
 	docker push $(2)
 endef
 
+.PHONY: help
 help:        ## Show this help.
              ##
 	@fgrep -h "##" $(MAKEFILE_LIST) | fgrep -v fgrep | sed -e 's/\\$$//' | sed -e 's/##//'
 
-clean:       ## Clean build time resources, primarily docker resources.
+.PHONY: clean
+clean:       ## Clean build time resources, primarily, unused docker images.
              ##
 	for i in `docker images | grep -E '^gmp/|.*/prometheus-engine/.*' | awk '{print $$3}' | uniq`; do docker image rm -f $$i; done
 
+.PHONY: lint
 lint:        ## Lint code.
              ##
 	@echo ">> linting code"
@@ -83,29 +86,27 @@ bin:         ## Build all go binaries from cmd/.
              ##
 bin: $(GOCMDS)
 
+.PHONY: regen
 regen:       ## Refresh autogenerated files and reformat code.
-             ## Use DRY_RUN=1 to only validate without generating changes.
+             ## Use CHECK=1 to only validate clean repo after run.
              ##
 regen:
-ifeq ($(DRY_RUN), 1)
+ifeq ($(CHECK), 1)
 	$(call docker_build, -f ./hack/Dockerfile --target hermetic -t gmp/hermetic \
 		--build-arg RUNCMD='./hack/presubmit.sh all diff' .)
 else
 	$(call docker_build, -f ./hack/Dockerfile --target sync -o . -t gmp/sync \
-		--build-arg RUNCMD='./hack/presubmit.sh' .)
+		--build-arg RUNCMD='./hack/presubmit.sh all' .)
 	rm -rf vendor && mv vendor.tmp vendor
 endif
 
-test:        ## Run all tests. Setting NO_DOCKER=1 writes real data to GCM API under PROJECT_ID environment variable.
-             ## Use GMP_CLUSTER, GMP_LOCATION to specify timeseries labels.
+.PHONY: test
+test:        ## Run unit tests. Setting NO_DOCKER=1 runs test on host machine.
              ##
-	@echo ">> running tests"
+	@echo ">> running unit tests"
 ifeq ($(NO_DOCKER), 1)
-	kubectl apply -f manifests/setup.yaml
-	kubectl apply -f cmd/operator/deploy/operator/01-priority-class.yaml
-	kubectl apply -f cmd/operator/deploy/operator/03-role.yaml
-	go test `go list ./... | grep -v operator/e2e | grep -v export/bench`
-	go test `go list ./... | grep operator/e2e` -args -project-id=${PROJECT_ID} -cluster=${GMP_CLUSTER} -location=${GMP_LOCATION}
+
+	go test `go list ./... | grep -v e2e | grep -v export/bench`
 else
 	# TODO(TheSpiritXIII): Temporary env variables part of `export.go` unit tests.
 	$(call docker_build, -f ./hack/Dockerfile --target sync -o . -t gmp/hermetic \
@@ -113,25 +114,51 @@ else
 	rm -rf vendor.tmp
 endif
 
-kindtest:    ## Run e2e test suite against fresh kind k8s cluster.
+.PHONY: e2e
+e2e:         ## Run e2e test suite against fresh kind k8s cluster. By default it
+             ## deploys and uses single-node kind cluster (left up after the tests). Setting
+             ## NO_DOCKER=1 execute tests from the host machine assuming kubectl
+             ## is already configured against the Kubernetes cluster of you choice.
              ##
-	@echo ">> building image"
+             ## Setting NO_DOCKER=1 also writes real data to GCM API under $PROJECT_ID
+             # environment variable. Use $GMP_CLUSTER, $GMP_LOCATION to specify
+             ## timeseries labels.
+             ##
+ifeq ($(NO_DOCKER), 1)
+	@echo ">> running e2e against your cluster"
+	go test `go list ./... | grep e2e` -args -project-id=${PROJECT_ID} -cluster=${GMP_CLUSTER} -location=${GMP_LOCATION}
+else
+	@echo ">> building kindtest image"
 	$(call docker_build, -f hack/Dockerfile --target kindtest -t gmp/kindtest .)
 	@echo ">> running container"
 # We lose some isolation by sharing the host network with the kind containers.
 # However, we avoid a gcloud-shell "Dockerception" and save on build times.
 	docker run --network host --rm -v $(DOCKER_VOLUME):/var/run/docker.sock gmp/kindtest ./hack/kind-test.sh
+endif
 
-presubmit:   ## Run all checks and tests before submitting a change 
-             ## Use DRY_RUN=1 to only validate without regenerating changes.
+.PHONY: presubmit
+presubmit:   ## Regenerate all resources, build all images and run all tests.
+             ## Steps from presubmit are validated on the CI, but feel free to
+             ## run it if you see CI failures related to regenerating resources
+             ## or when you want to do local check before submitting.
              ##
-presubmit: updateversions regen bin test kindtest
+             ## Use `CHECK=1` to fail the command if repo state is not clean
+             ## after presubmit (might require committing the changes).
+             ##
+presubmit: updateversions regen bin test e2e
 
+.PHONY: updateversions
 CURRENT_TAG = v0.7.1-gke.0
 CURRENT_PROM_TAG = v2.35.0-gmp.5-gke.0
 CURRENT_AM_TAG = v0.24.0-gmp.3-gke.1
 LABEL_API_VERSION = 0.7.1
 FILES_TO_UPDATE = $(shell find . -type f -name "*.yaml" ! -name "kube-state-metrics.yaml" ! -name "node-exporter.yaml")
+updateversions: ## Modifie all manifests, so it contains the expected versions.
+                ##
+                ## TODO(bwplotka): CI does not check updateversions--add that there.
+                ## Also, consider moving updateversion to hack/presubmit.sh for
+                ## consistency.
+                ##
 updateversions: $(SED)
 	@echo ">> Updating prometheus-engine images in manifests to $(CURRENT_TAG)"
 	@$(SED) -i -r 's#image: gke.gcr.io/prometheus-engine/(.*):.*#image: gke.gcr.io/prometheus-engine/\1:$(CURRENT_TAG)#g' $(FILES_TO_UPDATE) # This will match all, but Prom and AM will be fixed below.
