+++ README.md
@@ -27,31 +27,7 @@ see [GoogleCloudPlatform/prometheus](https://github.com/GoogleCloudPlatform/prom
 
 Images for this repo are regularly released [in the GKE release GCR](https://console.cloud.google.com/gcr/images/gke-release/global/prometheus-engine).
 
-## Build
-
-Run `make help` shows a list of candidate targets with documentation.
-
-Any go application in `./cmd/` with an associated `main.go`, e.g. `./cmd/operator/main.go`
-is a candidate for build by running:
-```bash
-make operator
-```
-
-- Running `make bin` will generate all the go binaries.
-  - Setting `NO_DOCKER=1` here will build all the binaries natively on the host machine.
-- Running `make test` will run unit and e2e tests.
-  - If `NO_DOCKER=1` is set, end-to-end tests will be run against the current
-  kubectl context. It is assumed the cluster has access to the GCM API.
-  Ensure `GMP_CLUSTER` and `GMP_LOCATION` are set, e.g.
-  ```bash
-  NO_DOCKER=1 GMP_CLUSTER=<my-cluster> GMP_LOCATION=<cluster-location> make test
-  ```
-- Running `make presubmit` will run various checks on the repo to ensure it is
-ready to submit a pull request. This includes testing, formatting,
-and regenerating files in-place.
-  - Setting `DRY_RUN=1` won't regenerate any files but will return a
-  non-zero exit code if the current changes differ from what would be. This
-  can be useful in running in CI workflows.
+## Development
 
 ### Dependencies
 
@@ -60,32 +36,105 @@ recommended:
 1. [`go`](https://golang.org/doc/install)
 2. [`gcloud`](https://cloud.google.com/sdk/docs/install)
 3. [`kubectl`](https://kubernetes.io/docs/tasks/tools/)
-  - Can also be installed via
-  ```
+
+Can be also installed via:
+
+  ```bash
   gcloud components install kubectl
   ```
 4. [`Docker`](https://docs.docker.com/get-docker/) with
-[`buildx`](https://docs.docker.com/build/architecture/#install-buildx) plugin
-  - Can also be run via
+   [`buildx`](https://docs.docker.com/build/architecture/#install-buildx) plugin.
+
+If you want to execute docker containers on remove machine you can run:
   ```bash
-  gcloud alpha cloud-shell ssh -- -nNT -L `pwd`/docker.sock:/var/run/docker.sock
+  gcloud alpha cloud-shell ssh --authorize-session -- -nNT -L `pwd`/docker.sock:/var/run/docker.sock
+  
   # Then in separate terminal.
   export DOCKER_HOST=unix://docker.sock
   ```
+
 5. For UI development or update (e.g. to resolve UI security issue), `npm` is
-required. See [pkg/ui documentation](pkg/ui/README.md) for details.
+   required. See [pkg/ui documentation](pkg/ui/README.md) for details.
+
+### Presubmit
+
+`make presubmit` regenerates all resources, builds all images and runs all tests.
+
+Steps from presubmit are validated on the CI, but feel free to run it if you see
+CI failures related to regenerating resources or when you want to do local check
+before submitting.
+
+Run `CHECK=1 make presubmit` to fail the command if repo state is not clean after
+presubmit (might require committing the changes).
+
+### Building
+
+Run `make help` shows a list of candidate targets with documentation.
+
+Any go application in `./cmd/` with an associated `main.go`, e.g. `./cmd/operator/main.go`
+is a candidate for build by running, for example:
+
+```bash
+make operator
+make frontend
+make rule-evaluator
+make example-app
+make config-reloader
+```
+
+Running `make bin` will build all of th[kind-test.sh](hack%2Fkind-test.sh)e above go binaries.
+  * Setting `NO_DOCKER=1` here will build all the binaries natively on the host machine.
 
 ### Testing
 
 #### Unit
 
-To run unit tests, use `go test`, your IDE or `make test`.
+To run unit tests locally, use `go test`, your IDE or `NO_DOCKER=1 make test`.
 
-#### e2e
+To run unit tests from docker container run `make test`
 
-To run e2e tests using kind cluster (created for you) use `make kind`.
+#### End-to-end tests
 
-The tests are available in ...
+Running `make e2e` will run e2e tests against Kubernetes cluster:
+  * By default, it run in hermetic docker container, downloads kind, recreates
+a single node kind cluster and runs [e2e](./e2e) tests against it.
+  * If `NO_DOCKER=1` is set, end-to-end tests will be run against the current
+    kubectl context. It is assumed the cluster has access to the GCM API.
+    Ensure `GMP_CLUSTER` and `GMP_LOCATION` are set, e.g.
+  ```bash
+  NO_DOCKER=1 GMP_CLUSTER=<my-cluster> GMP_LOCATION=<cluster-location> make e2e
+  ```
+
+In docker mode, to run single a test or debug a cluster during or after failed
+test, you can try entering shell of the `kindtest` container. Before doing so, 
+run `make e2e` to setup `kind` and start a cluster.
+
+To enter shell with kind Kubernetes context, (ensure your docker socket is on
+`/var/run/docker.sock`):
+
+```bash
+docker run --network host --rm -it \
+  -v /var/run/docker.sock:/var/run/docker.sock \
+  -v `pwd`/e2e:/build/e2e gmp/kindtest bash
+```
+
+To access kind Kubernetes (e.g. to list pods) run:
+
+```bash
+kind export kubeconfig
+kubectl get po
+```
+
+To execute a single test e.g. `TestAlertmanagerDefault` you can do (in `kindtest` shell):
+
+```bash
+kind export kubeconfig
+go test -v ./e2e -run "TestAlertmanagerDefault" -args -project-id=test-proj -cluster=test-cluster -location=test-loc -skip-gcm
+```
+
+Each test case is creating a separate set of namespaces e.g. `gmp-test-testalertmanagerdefault-20230714-120756` and `gmp-test-testalertmanagerdefault-20230714-120756-pub`, so to debug tests you
+have to ensure those namespaces are not cleaned. You can also provide time.Sleep in
+the place you want debug in.
 
 ##### Benchmarking
