+++ e2e/framework_test.go
@@ -12,14 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Package e2e contains tests that validate the behavior of gmp-operator against a cluster.
+// To make tests simple and fast, the test suite runs the operator internally. The CRDs
+// are expected to be installed out of band (along with the operator deployment itself in
+// a real world setup).
 package e2e
 
 import (
 	"context"
+	"flag"
 	"fmt"
 	"io/ioutil"
 	"os"
+	"os/signal"
 	"strings"
+	"syscall"
 	"testing"
 	"time"
 
@@ -31,6 +38,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/rest"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
@@ -41,15 +49,59 @@ import (
 var (
 	startTime    = time.Now().UTC()
 	globalLogger = zap.New(zap.Level(zapcore.Level(-1)))
+
+	kubeconfig        *rest.Config
+	projectID         string
+	cluster           string
+	location          string
+	skipGCM           bool
+	gcpServiceAccount string
 )
 
 func init() {
 	ctrl.SetLogger(globalLogger)
 }
 
+// TestMain injects custom flags and adds extra signal handling to ensure testing
+// namespaces are cleaned after tests were executed.
+func TestMain(m *testing.M) {
+	flag.StringVar(&projectID, "project-id", "", "The GCP project to write metrics to.")
+	flag.StringVar(&cluster, "cluster", "", "The name of the Kubernetes cluster that's tested against.")
+	flag.StringVar(&location, "location", "", "The location of the Kubernetes cluster that's tested against.")
+	flag.BoolVar(&skipGCM, "skip-gcm", false, "Skip validating GCM ingested points.")
+	flag.StringVar(&gcpServiceAccount, "gcp-service-account", "", "Path to GCP service account file for usage by deployed containers.")
+
+	flag.Parse()
+
+	var err error
+	kubeconfig, err = ctrl.GetConfig()
+	if err != nil {
+		fmt.Fprintln(os.Stderr, "Loading kubeconfig failed:", err)
+		os.Exit(1)
+	}
+
+	go func() {
+		os.Exit(m.Run())
+	}()
+
+	// If the process gets terminated by the user, the Go test package
+	// doesn't ensure that test cleanup functions are run.
+	// Deleting all namespaces ensures we don't leave anything behind regardless.
+	// Non-namespaced resources are owned by a namespace and thus cleaned up
+	// by Kubernetes' garbage collection.
+	term := make(chan os.Signal, 1)
+	signal.Notify(term, os.Interrupt, syscall.SIGTERM)
+
+	<-term
+	if err := cleanupAllNamespaces(context.Background()); err != nil {
+		fmt.Fprintln(os.Stderr, "Cleaning up namespaces failed:", err)
+		os.Exit(1)
+	}
+}
+
 // testContext manages shared state for a group of test. Test contexts are isolated
-// based on a unqiue namespace. This requires that no test affects or can be affected by
-// resources outside of the namespace managed by the context.
+// based on an unqiue namespace. This requires that no test affects or can be affected by
+// resources outside the namespace managed by the context.
 // The cluster must be left in a clean state after the cleanup handler completed successfully.
 type testContext struct {
 	*testing.T
