--- pkg/operator/e2e/alertmanager.yaml
@@ -1,1 +1,0 @@
-../../../cmd/operator/deploy/operator/12-alertmanager.yaml
