--- pkg/operator/e2e/collector.yaml
@@ -1,1 +1,0 @@
-../../../cmd/operator/deploy/operator/10-collector.yaml
