--- pkg/operator/e2e/rule-evaluator.yaml
@@ -1,1 +1,0 @@
-../../../cmd/operator/deploy/operator/11-rule-evaluator.yaml
